+++ cmd/deploy_helpers_test.go
@@ -234,7 +234,6 @@ func TestRunPrechecks(t *testing.T) {
 
 			info := lib.DeployInfo{TemplateRelativePath: "test"}
 			logObj := lib.DeploymentLog{}
-			outputsettings = settings.NewOutputSettings()
 
 			out := runPrechecks(&info, &logObj)
 
@@ -361,7 +360,6 @@ func TestPrepareDeployment(t *testing.T) {
 			viper.Set("logging.enabled", false)
 			viper.Set("templates.directory", "../examples/templates")
 			viper.Set("changeset.name-format", "changeset-$TIMESTAMP")
-			outputsettings = settings.NewOutputSettings()
 			deployFlags = DeployFlags{
 				StackName: tc.stackName,
 				Template:  tc.template,
@@ -449,7 +447,6 @@ func TestCreateAndShowChangeset(t *testing.T) {
 				Changeset: changeset,
 			}
 			logObj := lib.DeploymentLog{}
-			outputsettings = settings.NewOutputSettings()
 
 			cs := createAndShowChangeset(&info, config.AWSConfig{}, &logObj)
 
@@ -552,7 +549,6 @@ func TestConfirmAndDeployChangeset(t *testing.T) {
 				CreateChangeset: tc.createOnly,
 				NonInteractive:  tc.nonInteractive,
 			}
-			outputsettings = settings.NewOutputSettings()
 
 			result := confirmAndDeployChangeset(&lib.ChangesetInfo{}, &lib.DeployInfo{}, config.AWSConfig{})
 
@@ -647,7 +643,6 @@ func TestPrintDeploymentResults(t *testing.T) {
 			}
 
 			viper.Set("logging.enabled", false)
-			outputsettings = settings.NewOutputSettings()
 
 			info := lib.DeployInfo{IsNew: tc.isNew}
 			logObj := lib.DeploymentLog{}
@@ -699,7 +694,6 @@ func TestPrintDeploymentResults_WithOutputs(t *testing.T) {
 	}
 
 	viper.Set("logging.enabled", false)
-	outputsettings = settings.NewOutputSettings()
 
 	info := lib.DeployInfo{IsNew: true}
 	logObj := lib.DeploymentLog{}